// src/lib/types.ts
interface BaseObject {
  id: number
  name: string
}

export interface ShapeObject extends BaseObject {
  type: 'shape'
  widthM: number
  heightM: number
  color: string
}

export interface LineObject extends BaseObject {
  type: 'line'
  lengthM: number
  color: string
}

export interface MaskObject extends BaseObject {
  type: 'mask'
}

export interface BackgroundImageObject extends BaseObject {
  type: 'backgroundImage'
  imageData: string
}

export interface OverlayImageObject extends BaseObject {
  type: 'overlayImage'
  imageData: string
}

export type PlannerObject =
  | ShapeObject
  | LineObject
  | MaskObject
  | BackgroundImageObject
  | OverlayImageObject